profiles: refactor to separate sdk and exporter
