+++ exporters/otlp/profiles/src/main/java/io/opentelemetry/exporter/otlp/profiles/FunctionMarshaler.java
@@ -9,7 +9,7 @@ import io.opentelemetry.exporter.internal.marshal.MarshalerUtil;
 import io.opentelemetry.exporter.internal.marshal.MarshalerWithSize;
 import io.opentelemetry.exporter.internal.marshal.Serializer;
 import io.opentelemetry.proto.profiles.v1development.internal.Function;
-import io.opentelemetry.sdk.profiles.FunctionData;
+import io.opentelemetry.sdk.profiles.data.FunctionData;
 import java.io.IOException;
 import java.util.List;
 import java.util.function.Consumer;
