+++ exporters/otlp/profiles/src/main/java/io/opentelemetry/exporter/otlp/profiles/LocationMarshaler.java
@@ -9,7 +9,7 @@ import io.opentelemetry.exporter.internal.marshal.MarshalerUtil;
 import io.opentelemetry.exporter.internal.marshal.MarshalerWithSize;
 import io.opentelemetry.exporter.internal.marshal.Serializer;
 import io.opentelemetry.proto.profiles.v1development.internal.Location;
-import io.opentelemetry.sdk.profiles.LocationData;
+import io.opentelemetry.sdk.profiles.data.LocationData;
 import java.io.IOException;
 import java.util.List;
 import java.util.function.Consumer;
