+++ exporters/otlp/profiles/src/main/java/io/opentelemetry/exporter/otlp/profiles/NoopProfileExporter.java
@@ -6,8 +6,8 @@
 package io.opentelemetry.exporter.otlp.profiles;
 
 import io.opentelemetry.sdk.common.CompletableResultCode;
-import io.opentelemetry.sdk.profiles.ProfileData;
 import io.opentelemetry.sdk.profiles.ProfileExporter;
+import io.opentelemetry.sdk.profiles.data.ProfileData;
 import java.util.Collection;
 
 final class NoopProfileExporter implements ProfileExporter {
