+++ exporters/otlp/profiles/src/main/java/io/opentelemetry/exporter/otlp/profiles/ProfileMarshaler.java
@@ -9,7 +9,7 @@ import io.opentelemetry.exporter.internal.marshal.MarshalerUtil;
 import io.opentelemetry.exporter.internal.marshal.MarshalerWithSize;
 import io.opentelemetry.exporter.internal.marshal.Serializer;
 import io.opentelemetry.proto.profiles.v1development.internal.Profile;
-import io.opentelemetry.sdk.profiles.ProfileData;
+import io.opentelemetry.sdk.profiles.data.ProfileData;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.List;
