+++ exporters/otlp/profiles/src/main/java/io/opentelemetry/exporter/otlp/profiles/ProfilesDictionaryMarshaler.java
@@ -9,7 +9,7 @@ import io.opentelemetry.exporter.internal.marshal.MarshalerUtil;
 import io.opentelemetry.exporter.internal.marshal.MarshalerWithSize;
 import io.opentelemetry.exporter.internal.marshal.Serializer;
 import io.opentelemetry.proto.profiles.v1development.internal.ProfilesDictionary;
-import io.opentelemetry.sdk.profiles.ProfilesDictionaryData;
+import io.opentelemetry.sdk.profiles.data.ProfilesDictionaryData;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
