+++ exporters/otlp/profiles/src/main/java/io/opentelemetry/exporter/otlp/profiles/ProfilesRequestMarshaler.java
@@ -13,8 +13,8 @@ import io.opentelemetry.exporter.internal.marshal.MarshalerWithSize;
 import io.opentelemetry.exporter.internal.marshal.ProtoFieldInfo;
 import io.opentelemetry.exporter.internal.marshal.Serializer;
 import io.opentelemetry.proto.collector.profiles.v1development.internal.ExportProfilesServiceRequest;
-import io.opentelemetry.sdk.profiles.ProfileData;
-import io.opentelemetry.sdk.profiles.ProfilesDictionaryData;
+import io.opentelemetry.sdk.profiles.data.ProfileData;
+import io.opentelemetry.sdk.profiles.data.ProfilesDictionaryData;
 import io.opentelemetry.sdk.profiles.internal.data.ImmutableProfilesDictionaryData;
 import java.io.IOException;
 import java.util.Collection;
