+++ exporters/otlp/profiles/src/main/java/io/opentelemetry/exporter/otlp/profiles/SampleMarshaler.java
@@ -9,7 +9,7 @@ import io.opentelemetry.exporter.internal.marshal.MarshalerUtil;
 import io.opentelemetry.exporter.internal.marshal.MarshalerWithSize;
 import io.opentelemetry.exporter.internal.marshal.Serializer;
 import io.opentelemetry.proto.profiles.v1development.internal.Sample;
-import io.opentelemetry.sdk.profiles.SampleData;
+import io.opentelemetry.sdk.profiles.data.SampleData;
 import java.io.IOException;
 import java.util.List;
 import java.util.function.Consumer;
