+++ exporters/otlp/profiles/src/test/java/io/opentelemetry/exporter/otlp/profiles/FakeTelemetryUtil.java
@@ -7,8 +7,8 @@ package io.opentelemetry.exporter.otlp.profiles;
 
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
-import io.opentelemetry.sdk.profiles.ProfileData;
-import io.opentelemetry.sdk.profiles.ProfilesDictionaryData;
+import io.opentelemetry.sdk.profiles.data.ProfileData;
+import io.opentelemetry.sdk.profiles.data.ProfilesDictionaryData;
 import io.opentelemetry.sdk.profiles.internal.data.ImmutableProfileData;
 import io.opentelemetry.sdk.profiles.internal.data.ImmutableProfilesDictionaryData;
 import io.opentelemetry.sdk.profiles.internal.data.ImmutableValueTypeData;
