+++ exporters/otlp/profiles/src/test/java/io/opentelemetry/exporter/otlp/profiles/ProfilesRequestMarshalerTest.java
@@ -29,16 +29,16 @@ import io.opentelemetry.proto.profiles.v1development.ScopeProfiles;
 import io.opentelemetry.proto.profiles.v1development.Stack;
 import io.opentelemetry.proto.profiles.v1development.ValueType;
 import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
-import io.opentelemetry.sdk.profiles.FunctionData;
-import io.opentelemetry.sdk.profiles.KeyValueAndUnitData;
-import io.opentelemetry.sdk.profiles.LineData;
-import io.opentelemetry.sdk.profiles.LinkData;
-import io.opentelemetry.sdk.profiles.LocationData;
-import io.opentelemetry.sdk.profiles.MappingData;
-import io.opentelemetry.sdk.profiles.ProfileData;
-import io.opentelemetry.sdk.profiles.SampleData;
-import io.opentelemetry.sdk.profiles.StackData;
-import io.opentelemetry.sdk.profiles.ValueTypeData;
+import io.opentelemetry.sdk.profiles.data.FunctionData;
+import io.opentelemetry.sdk.profiles.data.KeyValueAndUnitData;
+import io.opentelemetry.sdk.profiles.data.LineData;
+import io.opentelemetry.sdk.profiles.data.LinkData;
+import io.opentelemetry.sdk.profiles.data.LocationData;
+import io.opentelemetry.sdk.profiles.data.MappingData;
+import io.opentelemetry.sdk.profiles.data.ProfileData;
+import io.opentelemetry.sdk.profiles.data.SampleData;
+import io.opentelemetry.sdk.profiles.data.StackData;
+import io.opentelemetry.sdk.profiles.data.ValueTypeData;
 import io.opentelemetry.sdk.profiles.internal.data.ImmutableFunctionData;
 import io.opentelemetry.sdk.profiles.internal.data.ImmutableKeyValueAndUnitData;
 import io.opentelemetry.sdk.profiles.internal.data.ImmutableLineData;
