+++ opentelemetry-sdk-profiles/build.gradle.kts
@@ -1,25 +1,12 @@
 plugins {
   id("otel.java-conventions")
 //  id("otel.publish-conventions")
+
+  id("otel.animalsniffer-conventions")
 }
 
 description = "OpenTelemetry - Profiles SDK"
-otelJava.moduleName.set("io.opentelemetry.sdkprofiles")
-
-tasks {
-  // this module uses the jdk.jfr.consumer API, which was backported into 1.8 but is '@since 9'
-  // and therefore a bit of a pain to get gradle to compile against...
-  compileJava {
-    sourceCompatibility = "1.8"
-    targetCompatibility = "1.8"
-    options.release.set(null as Int?)
-  }
-  compileTestJava {
-    sourceCompatibility = "1.8"
-    targetCompatibility = "1.8"
-    options.release.set(null as Int?)
-  }
-}
+otelJava.moduleName.set("io.opentelemetry.sdk.profiles")
 
 dependencies {
   api(project(":sdk:common"))
