+++ opentelemetry-sdk-profiles/gradle.properties
@@ -1,0 +1,1 @@
+otel.release=alpha
