+++ opentelemetry-sdk-profiles/src/main/java/io/opentelemetry/sdk/profiles/data/StackData.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.profiles;
+package io.opentelemetry.sdk.profiles.data;
 
 import java.util.List;
 import javax.annotation.concurrent.Immutable;
