+++ opentelemetry-sdk-profiles/src/main/java/io/opentelemetry/sdk/profiles/internal/data/ImmutableSampleData.java
@@ -6,7 +6,7 @@
 package io.opentelemetry.sdk.profiles.internal.data;
 
 import com.google.auto.value.AutoValue;
-import io.opentelemetry.sdk.profiles.SampleData;
+import io.opentelemetry.sdk.profiles.data.SampleData;
 import java.util.List;
 import javax.annotation.concurrent.Immutable;
