+++ settings.gradle.kts
@@ -56,6 +56,7 @@ include(":integration-tests:graal")
 include(":integration-tests:graal-incubating")
 include(":javadoc-crawler")
 include(":opencensus-shim")
+include(":opentelemetry-jfr-profiles-shim")
 include(":opentelemetry-sdk-profiles")
 include(":opentracing-shim")
 include(":perf-harness")
